save reorder

+++ metrics.Data.Common/Infrastructure/Entities/MessageCategory.cs
@@ -12,6 +12,6 @@ namespace metrics.Data.Common.Infrastructure.Entities
         [MaxLength(100)]
         public string? Color { get; set; }
         
-        public int? SortOrder { get; set; }
+        public double? SortOrder { get; set; }
     }
 }
